requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
pyodbc
requests
gunicorn #es el servidor web profesional para producción